moved interaction classes into ImagePanel class
